# docs/api/openapi.json
{
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "API documentation generated from YAML routing definitions",
    "title": "GOTRS API",
    "version": "1.0.0"
  },
  "openapi": "3.0.0",
  "paths": {
    "/admin/customer-companies": {
      "get": {
        "description": "Get list of all customer companies",
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": [
          {
            "BearerAuth": [
              "customer:read"
            ]
          }
        ],
        "summary": "List Customer Companies",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/customer-companies/new": {
      "get": {
        "description": "Create a new customer company",
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": [
          {
            "BearerAuth": [
              "customer:write"
            ]
          }
        ],
        "summary": "Create Customer Company",
        "tags": [
          "admin"
        ]
      },
      "post": {
        "description": "Create a new customer company",
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": [
          {
            "BearerAuth": [
              "customer:write"
            ]
          }
        ],
        "summary": "Create Customer Company",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/customer-companies/{id}/delete": {
      "post": {
        "description": "Soft delete (deactivate) a customer company",
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": [
          {
            "BearerAuth": [
              "customer:delete"
            ]
          }
        ],
        "summary": "Delete Customer Company",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/customer-companies/{id}/edit": {
      "get": {
        "description": "Edit customer company details",
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": [
          {
            "BearerAuth": [
              "customer:write"
            ]
          }
        ],
        "summary": "Edit Customer Company",
        "tags": [
          "admin"
        ]
      },
      "post": {
        "description": "Edit customer company details",
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": [
          {
            "BearerAuth": [
              "customer:write"
            ]
          }
        ],
        "summary": "Edit Customer Company",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/customer-companies/{id}/portal": {
      "get": {
        "description": "Customize the customer portal for a company",
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": [
          {
            "BearerAuth": [
              "portal:admin"
            ]
          }
        ],
        "summary": "Portal Customization",
        "tags": [
          "admin"
        ]
      },
      "post": {
        "description": "Customize the customer portal for a company",
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": [
          {
            "BearerAuth": [
              "portal:admin"
            ]
          }
        ],
        "summary": "Portal Customization",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/customer-companies/{id}/portal/logo": {
      "post": {
        "description": "Upload a custom logo for the customer portal",
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": [
          {
            "BearerAuth": [
              "portal:admin"
            ]
          }
        ],
        "summary": "Upload Portal Logo",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/customer-companies/{id}/services": {
      "get": {
        "description": "Manage services available to a company",
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": [
          {
            "BearerAuth": [
              "customer:write",
              "service:assign"
            ]
          }
        ],
        "summary": "Manage Company Services",
        "tags": [
          "admin"
        ]
      },
      "post": {
        "description": "Manage services available to a company",
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": [
          {
            "BearerAuth": [
              "customer:write",
              "service:assign"
            ]
          }
        ],
        "summary": "Manage Company Services",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/customer-companies/{id}/tickets": {
      "get": {
        "description": "Get all tickets for a company",
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": [
          {
            "BearerAuth": [
              "customer:read",
              "ticket:read"
            ]
          }
        ],
        "summary": "List Company Tickets",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/customer-companies/{id}/users": {
      "get": {
        "description": "Get all users belonging to a company",
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": [
          {
            "BearerAuth": [
              "customer:read"
            ]
          }
        ],
        "summary": "List Company Users",
        "tags": [
          "admin"
        ]
      }
    },
    "/agent/customers": {
      "get": {
        "description": "List and search customers",
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": [
          {
            "BearerAuth": [
              "customer:read"
            ]
          }
        ],
        "summary": "List Customers",
        "tags": [
          "agent"
        ]
      }
    },
    "/agent/customers/{id}": {
      "get": {
        "description": "View customer details and ticket history",
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": [
          {
            "BearerAuth": [
              "customer:read"
            ]
          }
        ],
        "summary": "View Customer",
        "tags": [
          "agent"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Development server",
      "url": "http://localhost:8080"
    }
  ]
}